cms version 1.2

+++ config_file.php
@@ -1,10 +1,10 @@
 <?php
 include "connectBasa.php"; // подключамся к базе
+	$id_name_table = htmlspecialchars($_POST['name_ad_hidden_menu']);
+	$id_link = htmlspecialchars($_POST['name_ad_hidden_id']);
 
 
 if (isset($_POST['sub_ad_save'])) {
-	$id_name_table = htmlspecialchars($_POST['name_ad_hidden_menu']);
-	$id_link = htmlspecialchars($_POST['name_ad_hidden_id']);
 	$name_links = htmlspecialchars($_POST['name_ad_links']);
 	$name_href_links = htmlspecialchars($_POST['name_ad_href']);
 	$name_links_class = htmlspecialchars($_POST['name_ad_class']);
@@ -20,14 +20,15 @@ if (isset($_POST['sub_ad_save'])) {
 if (isset($_POST['delet_mune'])) {
 	$name_menu = htmlspecialchars($_POST['hidden_name_menu']);
 	$delete_file = __DIR__. '/admin/'. $name_menu. '.php'; // путь к файлу
-	
+	$delet_file_users = __DIR__. '/tamplete_menu/'. $name_menu. '.php'; // путь к файлу
 
 	$filename = basename($delete_file, ".php"); // обрезаем все кроме названия файла
 	
 
 	$delet_table = $conn->query("DROP TABLE $filename;"); // удаляем таблицу
-	if (file_exists($delete_file)) { // проверяем существует ли файл
-		unlink($delete_file); // удаляем сам файл
+	if (file_exists($delete_file) and file_exists($delet_file_users)) { // проверяем существует ли файл
+		unlink($delete_file); // удаляем сам файл Admin
+		unlink($delet_file_users); // удаляем сам файл Users
 
 	}
 	
@@ -38,7 +39,13 @@ if (isset($_POST['delet_mune'])) {
 
 }
 
+	if (isset($_POST['sub_delete_links'])) {
+		$delete_link = $conn->query("DELETE FROM `$id_name_table` WHERE id = $id_link");
+		
+		echo '<meta http-equiv = "Refresh" content = "0; URL = index.php">';
+		exit();
 
+	}
